Adding generating only model and discriminator

+++ hooks/99_packageProcessor.js
@@ -8,21 +8,29 @@ module.exports = {
      * @param generator
      */
     'generate:after': generator => {
-        const sourcePath = generator.targetDir + '/src/main/java/';
-        const testPath = generator.targetDir + '/src/test/java/';
         let javaPackage = generator.templateParams['userJavaPackage'];
 
         javaPackage = javaPackage.replace(/\./g, '/');
 
         if (javaPackage !== 'com/asyncapi') {
+
+            const sourcePath = generator.targetDir + '/src/main/java/';
+
             const tmpSrc = tmp.dirSync();
             const tmpTest = tmp.dirSync();
-            fs.copySync(sourcePath + 'com/asyncapi', tmpSrc.name);
-            fs.copySync(testPath + 'com/asyncapi', tmpTest.name);
-            fs.removeSync(sourcePath + 'com');
-            fs.removeSync(testPath + 'com');
-            fs.copySync(tmpSrc.name, sourcePath + javaPackage);
-            fs.copySync(tmpTest.name, testPath + javaPackage);
+            fs.copySync(sourcePath + 'com/asyncapi', tmpSrc.name);            
+            fs.removeSync(sourcePath + 'com');            
+            fs.copySync(tmpSrc.name, sourcePath + javaPackage); 
+
+            // when onlyModel is generated than we don't have any test files 
+            let onlyModel = generator.templateParams['onlyModel'];
+            if(onlyModel === 'false') {
+              const testPath = generator.targetDir + '/src/test/java/';
+              fs.copySync(testPath + 'com/asyncapi', tmpTest.name);
+              fs.removeSync(testPath + 'com');
+              fs.copySync(tmpTest.name, testPath + javaPackage);
+            }
+
             tmp.setGracefulCleanup();
         }
     },
